+++ Rivr.Core/Models/Orders/Order.cs
@@ -81,7 +81,7 @@ public class Order
     /// <summary>
     /// The metadata of the order.
     /// </summary>
-    public Dictionary<string, string>? Metadata { get; set; }
+    public Dictionary<string, string> Metadata { get; set; } = new();
 
     /// <summary>
     /// The receipts of the order.